% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_eval.R
\name{matchLabel}
\alias{matchLabel}
\alias{calRand}
\alias{rSpecc}
\title{Calculate the rand-index of two label assignments}
\usage{
matchLabel(reference, label)

calRand(confmat)

rSpecc(x, centers, iter.max = 100, nstart = 10)
}
\arguments{
\item{reference}{true label assignment.}

\item{label}{estimated label assignment.}

\item{confmat}{a 2-dimensional confusion matrix.}

\item{x}{numeric matrix of data.}

\item{centers}{the number of clusters.}

\item{iter.max}{the maximum number of iterations allowed.}

\item{nstart}{how many random sets in the kmeans step should be chosen? default is 10.}
}
\value{
\item{Rand}{rand-index}
\item{adjRand}{adjusted rand-index}

\item{.Data}{A vector of integers indicating the cluster to which each point is allocated.}
\item{size}{The number of points in each cluster.}
\item{totss}{The total sum of squares.}
\item{withinss}{Vector of within-cluster sum of squares, one component per cluster.}
\item{tot.withinss}{Total within-cluster sum of squares, i.e. `sum(withinss)`.}
\item{betweenss}{The between-cluster sum of squares, i.e. `totss-tot.withinss`.}
}
\description{
`matchLabel( )` calculate the rand-index and adjusted rand-index of estimated label assignment
compared to true label assignment, which can be used to evaluate the performance of the estimated
label. The metric takes value between 0 and 1, and higher value indicates better performance.

`calRand( )` calculates the rand-index from a confusion matrix.

`rSpecc( )` a customized spectral clustering model.
}
