% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_mod.R, R/cluster_param.R, R/obj_logL.R
\name{heterogbcm}
\alias{heterogbcm}
\alias{obj_qc}
\alias{obj_qalpha}
\alias{obj_ppi}
\alias{obj_sigma}
\alias{obj_hlambda}
\alias{obj_hsigma}
\alias{obj_logL}
\title{Estimate the optimal posterior distribution of the data column labels.}
\usage{
heterogbcm(X, centers, tol, iter, iter_init = 3, labels, verbose = FALSE)

obj_qc(X, centers, ppi, sigma, qalpha, hlambda, hsigma)

obj_qalpha(X, centers, sigma, qc, hlambda, hsigma)

obj_ppi(centers, qc)

obj_sigma(centers, qalpha)

obj_hlambda(X, centers, qc, qalpha)

obj_hsigma(X, centers, qc, qalpha, hlambda)

obj_logL(X, centers, ppi, sigma, qc, qalpha, hlambda, hsigma)
}
\arguments{
\item{X}{matrix data.}

\item{centers}{An integer specifying the number of clusters.}

\item{tol}{numerical tolerance of the iteration updates.}

\item{iter}{number of iterations.}

\item{iter_init}{number of iterations of parameters initial estimation, default is 3.}

\item{labels}{A vector specifying the cluster labels of the columns of X.}

\item{verbose}{if TRUE, print iteration information.}

\item{ppi}{probability of multi-nulli distribution.}

\item{sigma}{group-correlation matrix.}

\item{qalpha}{posterior distribution of parameter vector alpha.}

\item{hlambda}{heterogeneous parameter vector Lambda.}

\item{hsigma}{heterogeneous parameter vector Sigma.}

\item{qc}{posterior distribution of labels.}
}
\value{
A list of values.
\item{sigma}{estimated optimal group-correlation matrix.}
\item{hlambda}{estimated optimal heterogeneous parameter Lambda.}
\item{hsigma}{estimated optimal heterogeneous parameter Sigma.}
\item{obj_logL_val}{vector of -logL from each iteration.}
\item{qc}{estimated optimal posterior distribution of the column labels.}
}
\description{
`heterogbcm( )` gives the optimal posterior distribution of the labels, which can be used
to derive the optimal label assignment of the data columns.
}
