% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_visual.R
\name{colMat_reorder}
\alias{colMat_reorder}
\alias{colMat_heatMap}
\alias{crossValid_func_adjR}
\title{Visulize the clustering results}
\usage{
colMat_reorder(affMatrix, centers, labels)

colMat_heatMap(
  affMatrix,
  centers,
  labels,
  margin = 0.5,
  midpoint = 0,
  limit = c(-1, 1),
  size = 0.2,
  legendName = "",
  title = ""
)

crossValid_func_adjR(x, centers, pt)
}
\arguments{
\item{affMatrix}{a square affinity matrix.}

\item{centers}{number of clusters.}

\item{labels}{label assignment of the columns.}

\item{margin}{a numeric value adjusting the position of group lines.}

\item{midpoint}{a numeric value of the midpoint of heatmap range.}

\item{limit}{a numeric range of the values.}

\item{size}{size of the groupline.}

\item{legendName}{a character name of the legend.}

\item{title}{a character name of the heatmap.}

\item{x}{an input matrix.}

\item{pt}{partition times.}
}
\value{
\item{matReorder}{affinity matrix reordered by groups.}
\item{groupLens}{a vector of group lengths.}
}
\description{
`colMat_reorder( )` reorder the columns of affinity matrix by specified groups.

`colMat_heatMap( )` plots the heatmap of affMatrix ordered by groups.

`crossValid( )` 2-fold crossValidation selects the number of clusters.
}
