% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Generate simulation data and parameters}
\usage{
data_gen(n, p, centers, mu, sigma, labels, size, hpara_func)
}
\arguments{
\item{n}{An integer specifying the number of observations (rows) of data X.}

\item{p}{An integer specifying the number of variables (columns) of data X.}

\item{centers}{An integer specifying the number of clusters.}

\item{mu}{A vector of size `p` specifying the mean of the multivariate normal distribution of Alpha.}

\item{sigma}{A matrix of size `p x p` specifying the variance matrix of the 
multivariate normal distribution of Alpha.}

\item{labels}{A vector of size `p` specifying the cluster labels of the variables.}

\item{size}{An integer specifying the number of simulated data sets.}

\item{hpara_func}{A list of size 2 specifying the function of generating random numbers for
parameter Lambda and Sigma.}
}
\value{
A list of lists, each sub-list contains a set of data X, parameter Alpha, Lambda 
and Sigma.
}
\description{
`data_gen( )` generates simulation data and parameters for implementation HBCM model
}
\examples{
n <- 100
p <- 2
centers <- 2
mu <- rep(0, p)
sigma <- matrix(c(1,0,0,1), nrow=p)
labels <- c(1,2)
size <- 5
hpara_func <- list(
  lambda_func = function(p) stats::rnorm(p, 0, 1),
  sigma_func = function(p) stats::rchisq(p, 2) + 1
)
data_list <- data_gen(n, p, centers, mu, sigma, labels, size, hpara_func)
}
