% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gene_group}
\alias{data_gene_group}
\title{Transcript Factor (TF) regulators of yeast genes.}
\format{
A data frame with three variables: \code{Probe}, \code{Gene}, \code{TF_regulator}.
}
\usage{
data_gene_group
}
\description{
The raw data is downloaded from the supplementary file of paper https://www.tandfonline.com/doi/suppl/10.1080/15384101.2019.1570655?scroll=top.
OR, see github https://github.com/xiangli2pro/hbcm/tree/main/inst/extdata/ to access the raw data.
Check github https://github.com/xiangli2pro/hbcm/tree/main/data-raw/yeast_gene_data.R to see how data is processed.
}
\keyword{datasets}
