% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_gene_sample}
\alias{data_gene_sample}
\title{Yeast gene expression from the experiment.}
\format{
A data frame with 241 variables (genes).
}
\usage{
data_gene_sample
}
\description{
The raw data is downloaded the paper https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75694
OR, see github https://github.com/xiangli2pro/hbcm/tree/main/inst/extdata/ to access the raw data.
Check github https://github.com/xiangli2pro/hbcm/tree/main/data-raw/yeast_gene_data.R to see how data is processed.
}
\keyword{datasets}
