% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{heterogbcm_expr1}
\alias{heterogbcm_expr1}
\title{Estimate the optimal posterior distribution of the data column labels.}
\usage{
heterogbcm_expr1(
  x,
  centers,
  tol,
  iter,
  labels,
  verbose = FALSE,
  init_hlambda,
  init_hsigma
)
}
\arguments{
\item{x}{a numeric matrix data.}

\item{centers}{an integer specifying the number of clusters.}

\item{tol}{numerical tolerance of the iteration updates.}

\item{iter}{number of iterations.}

\item{labels}{a vector specifying the cluster labels of the columns of x.}

\item{verbose}{if TRUE, print iteration information.}

\item{init_hlambda}{initial values for parameter vector Lambda.}

\item{init_hsigma}{initial values for parameter vector Sigma.}
}
\value{
A list of values.
\item{omega}{estimated optimal group-correlation matrix.}
\item{hlambda}{estimated optimal heterogeneous parameter Lambda.}
\item{hsigma}{estimated optimal heterogeneous parameter Sigma.}
\item{obj_logL_val}{vector of -logL from each iteration.}
\item{qc}{estimated optimal posterior distribution of the column labels.}
\item{cluster}{a vector of integers (from 1:k) indicating the cluster to which each column is allocated.}
}
\description{
`heterogbcm_expr1( )` gives the optimal posterior distribution of the labels, use random initial values for hlambda and hsigma
}
