% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbcm.R, R/obj_logL.R
\name{hbcm}
\alias{hbcm}
\alias{obj_logL}
\title{Estimate the optimal posterior distribution of the data column labels.}
\usage{
hbcm(X, centers, tol, iter, labels, verbose = FALSE)

obj_logL(X, centers, ppi, sigma, qc, qalpha, hlambda, hsigma)
}
\arguments{
\item{X}{matrix data.}

\item{centers}{An integer specifying the number of clusters.}

\item{tol}{numerical tolerance of the iteration updates.}

\item{iter}{number of iterations.}

\item{labels}{A vector specifying the cluster labels of the columns of X.}

\item{verbose}{if TRUE, print parameters estimation on each iteration.}

\item{ppi}{probability of multi-nulli distribution.}

\item{sigma}{group-correlation matrix.}

\item{qc}{posterior distribution of labels.}

\item{qalpha}{distribution of parameter alpha.}

\item{hlambda}{heterogeneous parameter Lambda.}

\item{hsigma}{heterogeneous parameter Sigma.}
}
\value{
A list of values.
\item{sigma}{estimated optimal group-correlation matrix.}
\item{hlambda}{estimated optimal heterogeneous parameter Lambda.}
\item{hsigma}{estimated optimal heterogeneous parameter Sigma.}
\item{obj_logL_val}{-logL after each iteration.}
\item{qc}{estimated optimal posterior distribution of the column labels.}
}
\description{
`hbcm( )` gives the optimal posterior distribution of the labels, which can be used
to derive the optimal label assignment of the data columns.
}
