% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{sample_gen}
\alias{sample_gen}
\title{Title}
\usage{
sample_gen(n, p, mu, sigma, labels, hpara_func)
}
\arguments{
\item{n}{}

\item{p}{}

\item{mu}{}

\item{sigma}{}

\item{labels}{}

\item{hpara_func}{}
}
\value{

}
\description{
Title
}
