% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Generate simulation data and parameters}
\usage{
data_gen(n, p, centers, mu, omega, labels, size, hparam_func)
}
\arguments{
\item{n}{an integer specifying the number of observations (rows) of data x.}

\item{p}{an integer specifying the number of variables (columns) of data x.}

\item{centers}{an integer specifying the number of clusters.}

\item{mu}{a vector of size `centers` specifying the mean vector of the 
multivariate normal distribution of alpha.}

\item{omega}{a matrix of size `centers x centers` specifying the covariance matrix of the
multivariate normal distribution of alpha.}

\item{labels}{a vector of size `p` specifying the cluster labels of the variables.}

\item{size}{an integer specifying the number of simulation data sets.}

\item{hparam_func}{a list of size two specifying the function of generating random numbers of
parameter Lambda and Sigma.}
}
\value{
A list of lists.
\item{x_list}{a list of data x.}
\item{alpha_list}{a list of alpha.}
\item{hlambda_list}{a list of Lambda.}
\item{hsigma_list}{a list of Sigma.}
}
\description{
`data_gen( )` generates matrix data (x) and parameters (alpha, Lambda, Sigma) for experimenting
with the HBCM model.
}
\examples{
n <- 500
p <- 500
centers <- 3

mu <- rep(0, centers)

off_diag <- 0.5
omega <- diag(rep(1, centers))
for (i in 1:centers) {
 for (j in 1:centers) {
   if (i!=j){
     omega[i,j] = off_diag
   } 
 }
}

ppi <- rep(1/centers, centers)
labels <- sample(c(1:centers), size = p, replace = TRUE, prob = ppi) 

size <- 5
hparam_func <- list(
  lambda_func = function(p) stats::rnorm(p, 0, 1),
  sigma_func = function(p) stats::rchisq(p, 2) + 1
)
data_list <- data_gen(n, p, centers, mu, omega, labels, size, hparam_func)
}
