% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_functions.R
\name{heterogbcm_hparam}
\alias{heterogbcm_hparam}
\alias{heterogbcm_constant_hlambda_hsigma}
\alias{heterogbcm_logL}
\alias{heterogbcm_iterStep}
\alias{heterogbcm_noInitLabel}
\alias{heterogbcm_qcDiscrete}
\title{Estimate the optimal posterior distribution of the data column labels.}
\usage{
heterogbcm_hparam(
  x,
  centers,
  tol,
  iter,
  labels,
  verbose = FALSE,
  init_hlambda,
  init_hsigma
)

heterogbcm_constant_hlambda_hsigma(
  x,
  centers,
  tol,
  iter,
  iter_init = 3,
  labels,
  verbose = FALSE
)

heterogbcm_logL(x, centers, tol, iter, iter_init = 3, labels, verbose = FALSE)

heterogbcm_iterStep(
  x,
  centers,
  tol,
  iter,
  iter_init = 3,
  labels,
  verbose = FALSE
)

heterogbcm_noInitLabel(
  x,
  centers,
  tol,
  iter,
  iter_init = 3,
  labels = NA,
  verbose = FALSE
)

heterogbcm_qcDiscrete(
  x,
  centers,
  tol,
  iter,
  iter_init = 3,
  labels,
  verbose = FALSE
)
}
\arguments{
\item{x}{a numeric matrix data.}

\item{centers}{an integer specifying the number of clusters.}

\item{tol}{numerical tolerance of the iteration updates.}

\item{iter}{number of iterations.}

\item{labels}{a vector specifying the cluster labels of the columns of x.}

\item{verbose}{if TRUE, print iteration information.}

\item{init_hlambda}{initial values for parameter vector Lambda.}

\item{init_hsigma}{initial values for parameter vector Sigma.}

\item{iter_init}{iteration times of initial parameter estimation}
}
\value{
A list of values.
\item{omega}{estimated optimal group-correlation matrix.}
\item{hlambda}{estimated optimal heterogeneous parameter Lambda.}
\item{hsigma}{estimated optimal heterogeneous parameter Sigma.}
\item{obj_logL_val}{vector of -logL from each iteration.}
\item{qc}{estimated optimal posterior distribution of the column labels.}
\item{cluster}{a vector of integers (from 1:k) indicating the cluster to which each column is allocated.}
}
\description{
`heterogbcm_hparam( )` gives the optimal posterior distribution of the labels, use random initial values for hlambda and hsigma.

`heterogbcm_constant_hlambda_hsigma( )` experiment function to keep hlambda and hsigma constant across simulation

`heterogbcm_logL( )` experiment function to compare the initial logL and last logL.

`heterogbcm_iterStep( )` experiment function with different update algorithm
first update q2 and parameters Theta, then update q1 and parameters Theta.

`heterogbcm_noInitLabel( )` experiment function with no initial guess of the labels.

`heterogbcm_qcDiscrete( )` experiment function to update qc with discrete values.
}
