% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_init.R, R/param_init_auxfunc.R
\name{init_hparam}
\alias{init_hparam}
\alias{init_hparam0}
\alias{init_sigma}
\alias{obj_init_qalpha}
\alias{obj_qalpha_logL}
\alias{obj_init_hlambda}
\alias{obj_init_hsigma}
\title{Initial estimation of heterogeneous parameters Lambda and Sigma}
\usage{
init_hparam(x, centers, labels, tol, iter, verbose = FALSE)

init_hparam0(x, tol, iter, verbose = FALSE)

init_sigma(x, centers, labels, hlambda, hsigma)

obj_init_qalpha(x, hlambda, hsigma)

obj_qalpha_logL(x, qalpha, hlambda, hsigma)

obj_init_hlambda(x, qalpha)

obj_init_hsigma(x, qalpha, hlambda)
}
\arguments{
\item{x}{a numeric matrix data.}

\item{centers}{an integer specifying the number of clusters.}

\item{labels}{a vector specifying the cluster labels of the columns of x.}

\item{tol}{numerical tolerance of the iteration updates.}

\item{iter}{number of iterations.}

\item{verbose}{if TRUE, print iteration information.}

\item{hlambda}{heterogeneous parameter vector Lambda.}

\item{hsigma}{heterogeneous parameter vector Sigma.}

\item{qalpha}{distribution of parameter vector alpha.}
}
\value{
A list of values.
\item{hlambda}{optimal initial estimation of Lambda.}
\item{hsigma}{optimal initial estimation of Sigma.}
}
\description{
`init_hparam( )` gives the initial estimation of parameters Lambda and Sigma.

`init_sigma( )` gives the initial estimation of group-correlation matrix sigma.
}
