% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_init.R, R/param_init_auxfunc.R
\name{init_hparam}
\alias{init_hparam}
\alias{init_hparam0}
\alias{obj_init_qalpha}
\alias{obj_qalpha_logL}
\alias{obj_init_hlambda}
\alias{obj_init_hsigma}
\alias{init_sigma}
\title{Initial estimation of heterogeneous parameters Lambda and Sigma}
\usage{
init_hparam(X, centers, labels, tol, iter, verbose = FALSE)

init_hparam0(X, tol, iter, verbose = FALSE)

obj_init_qalpha(X, hlambda, hsigma)

obj_qalpha_logL(X, qalpha, hlambda, hsigma)

obj_init_hlambda(X, qalpha)

obj_init_hsigma(X, qalpha, hlambda)

init_sigma(X, centers, labels, hlambda, hsigma)
}
\arguments{
\item{X}{matrix data.}

\item{centers}{An integer specifying the number of clusters.}

\item{labels}{A vector specifying the cluster labels of the columns of X.}

\item{tol}{numerical tolerance of the iteration updates.}

\item{iter}{number of iterations.}

\item{verbose}{if TRUE, print parameters estimation on each iteration.}

\item{hlambda}{estimated heterogeneous parameter Lambda.}

\item{hsigma}{estimated heterogeneous parameter Sigma.}

\item{qalpha}{estimated distribution of parameter alpha.}
}
\value{
A list containing a vector of Lambda estimations and a vector of Sigma estimations
}
\description{
`init_hparam( )` gives the initial estimation of parameters Lambda and Sigma.

`init_hparam0( )` gives the initial estimation of parameters Lambda and Sigma when centers = 1.

`init_sigma( )` gives the initial estimation of group-correlation matrix
}
